Fix mypy type errors in backtest models and processor
Wrap float ** float expressions with float() to resolve no-any-return
errors in Benchmark and FutureTrade annualized_pct properties. Replace
boolean is_date_only variable with an isinstance chain in processor.run
so mypy can narrow date | datetime | None to datetime | None correctly.

+++ turtle/backtest/models.py
@@ -52,7 +52,7 @@ class Benchmark:
         days = (self.exit_date - self.entry_date).days
         if days <= 0:
             return self.return_pct
-        return ((1 + self.return_pct / 100.0) ** (365.0 / days) - 1) * 100.0
+        return float(((1 + self.return_pct / 100.0) ** (365.0 / days) - 1) * 100.0)
 
 
 @dataclass
@@ -123,7 +123,7 @@ class FutureTrade:
         holding_days = self.holding_days
         if holding_days <= 0:
             return self.realized_pct
-        return ((self.exit.price / self.entry.price) ** (365.0 / holding_days) - 1) * 100.0
+        return float(((self.exit.price / self.entry.price) ** (365.0 / holding_days) - 1) * 100.0)
 
     @property
     def exit_reason(self) -> str:
